requirement.txt
python3
pillow
numpy
networkx